autoload/grappele.vim: Only try to recall if a location was saved
Previously, if `gG` (the recall mapping) was pressed before previously
having pressed a `G` command, the following error would appear:

    Error detected while processing function grappele#Recall:
    line    1:
    E121: Undefined variable: s:line
    E116: Invalid arguments for function grappele#Grappele

This happened because no previous `G` location had been saved.

Displaying an error here makes for a bad user experience. Instead if
this happens, we should just do nothing. To prevent the error, check for
the existence of the `s:line` variable before trying to use it.